bridge: refactor out broadcastSignature to prepare for injection path
[ghstack-poisoned]
